pyproject.toml
[tool.poetry]
name = "scistream-proto"
version = "0.1.0"
description = ""
authors = ["Flavio Castro <fcastro@anl.gov>"]
readme = "README.md"
packages = [{include = "scistream_proto"}]

[tool.poetry.dependencies]
python = "^3.9"
optparse-pretty = "^0.1.1"
grpcio = "^1.51.3"
grpcio-tools = "^1.51.3"
grpclib = "^0.4.3"
click = "^8.1.3"
fire = "^0.5.0"
grpcio-testing = "^1.53.0"
pytest-timeout = "^2.1.0"
protobuf = "^4.22.3"
pyzmq = "^25.0.2"
globus-cli = "^3.15.0"
docker = "^6.1.3"
jinja2 = "^3.1.2"

[tool.poetry.group.test.dependencies]
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
